Moving API code to controller.
