+++ docker-compose.yml
@@ -7,7 +7,7 @@ version: '3.7'
 services:
 
   app:
-    image: registry.lts.harvard.edu/lts/mps-embed:0.0.4
+    image: registry.lts.harvard.edu/lts/mps-embed:0.0.5
     build:
       context: ./
       dockerfile: Dockerfile
